[module productList] rename to product
