+++ src/containers/Product/Product.js
@@ -1,6 +1,6 @@
 import { connect } from 'react-redux';
 import Product from '../../components/Product/Product';
-import { deleteProduct } from '../../redux/modules/productList';
+import { deleteProduct } from '../../redux/modules/product';
 
 const mapDispatchToProps = dispatch => ({
   delete: id => dispatch(deleteProduct(id)),
