+++ src/redux/modules/reducer.js
@@ -1,6 +1,6 @@
 import { combineReducers } from 'redux';
 import { routerReducer } from 'react-router-redux';
-import products from './productList';
+import products from './product';
 
 export default combineReducers({
   routing: routerReducer,
